Remove test for double-encoded JSON issue, add an image, add npm build script
Double-encoded JSON issue does not occur in the real application.
It was an issue with the test method. Tried to set up testing with
real JSON files but it's quite hard as it needs to be async.
Added value of the test is questionable. Won't catch bugs. Moving on!

Added an image just to figure out the syntax.

npm build script avoids some typing.